#4646 Merge in properly after initial merge
