+++ test2/faces22/cdiInitDestroyEvent/src/main/java/com/sun/faces/test/javaee7/cdiinitdestroyevent/ViewScopedBean.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 1997, 2020 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 1997, 2021 Oracle and/or its affiliates. All rights reserved.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License v. 2.0, which is available at
@@ -17,19 +17,35 @@
 package com.sun.faces.test.javaee7.cdiinitdestroyevent;
 
 import java.io.Serializable;
-
+import jakarta.annotation.PostConstruct;
+import jakarta.annotation.PreDestroy;
 import jakarta.faces.view.ViewScoped;
 import jakarta.inject.Named;
 
 @Named
 @ViewScoped
 public class ViewScopedBean implements Serializable {
-
+    
     private static final long serialVersionUID = -6301337067401894253L;
-
+    private static int counter = 0;
+    
     private String value = "My View Scoped Value";
 
     public String getValue() {
         return value;
+    }
+
+    public int getCounter() {
+        return counter;
+    }
+
+    @PostConstruct
+    public void postConstruct() {
+        counter++;
+    }
+
+    @PreDestroy
+    public void preDestroy() {
+        counter--;
     }
 }
