+++ test2/faces22/cdiInitDestroyEvent/src/main/webapp/issue4646.xhtml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!--
+
+    Copyright (c) 1997, 2019 Oracle and/or its affiliates. All rights reserved.
+
+    This program and the accompanying materials are made available under the
+    terms of the Eclipse Public License v. 2.0, which is available at
+    http://www.eclipse.org/legal/epl-2.0.
+
+    This Source Code may also be made available under the following Secondary
+    Licenses when the conditions for such availability set forth in the
+    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
+    version 2 with the GNU Classpath Exception, which is available at
+    https://www.gnu.org/software/classpath/license.html.
+
+    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+
+-->
+
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml"
+    xmlns:h="http://java.sun.com/jsf/html">
+    <head>
+        <title></title>
+    </head>
+    <body>
+        <h:form prependId="false">
+            <h:outputLabel id="counterMessage" value="#{viewScopedBean.counter}" />
+            <h:commandButton value="submit" />
+    
+            <p>
+                <h:commandButton id="invalidateSession" value="invalidate session"
+                    actionListener="#{appBean.invalidateSession}" action="page02?faces-redirect=true" />
+            </p>
+        </h:form>
+    </body>
+</html>
